Extract RulerPanel
Separates Ruler from RulerPanel to reduce unnecessary importing.
